backend/requirements.txt
Werkzeug==3.0.4
Flask==2.3.3
mysql-connector-python==8.4.0
flask-cors==5.0.0